+++ Jellyfin.Plugin.Edl/Configuration/PluginConfiguration.cs
@@ -12,6 +12,7 @@ public class PluginConfiguration : BasePluginConfiguration
     /// </summary>
     public PluginConfiguration()
     {
+        UnknownEdlAction = EdlAction.None;
         IntroEdlAction = EdlAction.None;
         OutroEdlAction = EdlAction.None;
         PreviewEdlAction = EdlAction.None;
@@ -20,9 +21,14 @@ public class PluginConfiguration : BasePluginConfiguration
     }
 
     /// <summary>
+    /// Gets or sets an Unknown Action option.
+    /// </summary>
+    public EdlAction UnknownEdlAction { get; set; }
+
+    /// <summary>
     /// Gets or sets an Intro Action option.
     /// </summary>
-    public EdlAction IntroEdlAction { get; set; }
+    public EdlAction IntroEdlAction { get; set; } = EdlAction.Cut;
 
     /// <summary>
     /// Gets or sets an Outro Action option.
@@ -42,7 +48,7 @@ public class PluginConfiguration : BasePluginConfiguration
     /// <summary>
     /// Gets or sets an Commercial Action option.
     /// </summary>
-    public EdlAction CommercialEdlAction { get; set; }
+    public EdlAction CommercialEdlAction { get; set; } = EdlAction.CommercialBreak;
 
     /// <summary>
     /// Gets or sets a value indicating whether to overwrite existing edl files. Which keeps the file in sync with media segment edits.
@@ -53,4 +59,19 @@ public class PluginConfiguration : BasePluginConfiguration
     /// Gets or sets the max degree of parallelism used when creating edl files.
     /// </summary>
     public int MaxParallelism { get; set; } = 2;
+
+    /// <summary>
+    /// Gets or sets the comma separated list of library names to analyze. If empty, all libraries will be analyzed.
+    /// </summary>
+    public string SelectedLibraries { get; set; } = string.Empty;
+
+    /// <summary>
+    /// Gets or sets the comma separated list of tv shows and seasons to skip the analyze. Format: "My Show;S01;S02, Another Show".
+    /// </summary>
+    public string SkippedTvShows { get; set; } = string.Empty;
+
+    /// <summary>
+    /// Gets or sets the comma separated list of movies to skip the analyze.".
+    /// </summary>
+    public string SkippedMovies { get; set; } = string.Empty;
 }